+++ Library/Aphid/UI/SourceListView.js
@@ -98,6 +98,7 @@
 Aphid.UI.SourceListView = Class.create(Aphid.UI.ListView, {
 
   displayName: "Aphid.UI.SourceListView",
+  tagName: "ul",
 
   viewDidLoad: function($super)
   {